+++ sw.js
@@ -19,7 +19,7 @@
  */
 'use strict';
 
-var VERSION = 'v112';
+var VERSION = 'v113';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;